+++ docs/source/conf.py
@@ -10,7 +10,7 @@ import sys
 
 sys.path.insert(0, os.path.abspath(".."))
 
-__version__ = "2025.07.02.1"
+__version__ = "2025.07.03.1"
 
 project = "django-blocknote"
 copyright = "2025, Ryan Sevelj & Mark Sevelj"